% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_polynomial.R
\name{basis.poly}
\alias{basis.poly}
\title{Internal Function: Generate a polynomial basis matrix of order r.}
\usage{
basis.poly(n, s, e, r)
}
\arguments{
\item{n}{An \code{integer} scalar of sample size.}

\item{s}{An \code{integer} scalar of starting index.}

\item{e}{An \code{integer} scalar of ending index.}

\item{r}{An \code{integer} scalar representing the order of polynomials.}
}
\value{
A (e-s+1)-by-(r+1) \code{numeric} basis matrix with (i,j)-th entry being \eqn{[(s+i-1)/n]^j}.
}
\description{
Internal Function: Generate a polynomial basis matrix of order r.
}
\examples{
poly_mat = basis.poly(100, 1, 50, 2)
}
\author{
Haotian Xu
}
