% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean_robust.R
\name{BD_U}
\alias{BD_U}
\title{Backward detection with a robust bootstrap change point test using U-statistics.}
\usage{
BD_U(y, M, B = 100, inter = NULL, inter_ini = TRUE)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{M}{An \code{integer} scalar representing initial block size of the backward detection algorithm.}

\item{B}{An \code{integer} scalar representing the number of bootstrapped samples.}

\item{inter}{A nuisance parameter.}

\item{inter_ini}{A nuisance parameter.}
}
\value{
An \code{numeric} vector of estimated change point locations.
}
\description{
Perform the backward detection method with a robust bootstrap change point test using U-statistics on univariate data.
}
\references{
Yu and Chen (2019) <arXiv:1904.03372>.
}
\author{
Mengchu Li
}
