% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoints.R
\name{binary_segmentation}
\alias{binary_segmentation}
\title{Estimates multiple change-points by binary segmentation}
\usage{
binary_segmentation(data, bbmod_init_vals, cp_method, bbmod_method, bbmod_ll,
  thresh = 0, buff = 100, cp_method_params = list(),
  bbmod_method_params = list(), bbmod_ll_params = list())
}
\arguments{
\item{data}{matrix of actual data.}

\item{bbmod_init_vals}{initial estimates for arbitrary bbmod values}

\item{cp_method}{single change-point method used}

\item{bbmod_method}{function corresponding to black box model itself}

\item{bbmod_ll}{function returning ll or comparable cost function for
black box model}

\item{thresh}{stopping threshold for cost comparison}

\item{buff}{distance from edge of sample to be maintained}

\item{cp_method_params}{list of additional parameters for cp_method}

\item{bbmod_method_params}{list of additional parameters for bbmod_method}

\item{bbmod_ll_params}{list of additional parameters for bbmod_ll}
}
\value{
List containing estimated change-points and thetas
}
\description{
Estimates multiple change-points by binary segmentations
             using provided single-change-point method to search each
             partition
}
\author{
Leland Bybee \email{leland.bybee@gmail.com}
}
