% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoints.R
\name{brute_force}
\alias{brute_force}
\title{Estimates a single change-point by searching all change-points}
\usage{
brute_force(data, bbmod_init_vals, bbmod_method, bbmod_ll, buff = 100,
  bbmod_method_params = list(), bbmod_ll_params = list())
}
\arguments{
\item{data}{matrix of actual data.}

\item{bbmod_init_vals}{initial estimates for arbitrary bbmod values}

\item{bbmod_method}{function corresponding to black box model itself}

\item{bbmod_ll}{function returning ll or comparable cost function for
black box model}

\item{buff}{distance from edge of sample to be maintained}

\item{bbmod_method_params}{list of additional parameters for bbmod_method}

\item{bbmod_ll_params}{list of additional parameters for bbmod_ll}
}
\value{
List containing estimated change-point and thetas
}
\description{
Estimates a single change-point by search all change-points
}
\author{
Leland Bybee \email{leland.bybee@gmail.com}
}
