% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD_covariance.R
\name{BS.cov}
\alias{BS.cov}
\title{Binary Segmentation for covariance change points detection through Operator Norm.}
\usage{
BS.cov(X, s, e, level = 0, ...)
}
\arguments{
\item{X}{A \code{numeric} matrix of observations with with horizontal axis being time, and vertical axis being dimensions.}

\item{s}{A \code{integer} scalar of starting index.}

\item{e}{A \code{integer} scalar of ending index.}

\item{level}{A parameter using to track the level a change point is detected. Should be fixed as 0.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item S           A vector of estimated changepoints (sorted in strictly increasing order).
 \item Dval        A vector of values of CUSUM statistic based on KS distance.
 \item Level       A vector representing the levels at which each change point is detected.
 \item Parent      A matrix with the starting indices on the first row and the ending indices on the second row.
 \item ...         Additional parameters.
}
}
\description{
TO DO
}
\examples{
p = 10
A1 = gen.cov.mat(p, 1, 2)
A2 = gen.cov.mat(p, 2, 1)
A3 = gen.cov.mat(p, 3, 2)
X = cbind(t(mvrnorm(100, mu = rep(0, p), A1)), t(mvrnorm(150, mu = rep(0, p), A2)), t(mvrnorm(200, mu = rep(0, p), A3)))
temp = BS.cov(X, 1, 450)
BS.threshold(temp, 10)
}
\author{
Haotian Xu
}
