% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{BS.network}
\alias{BS.network}
\title{Binary segmentation for network change points detection.}
\usage{
BS.network(data_mat1, data_mat2, s, e, delta, level = 0, ...)
}
\arguments{
\item{data_mat1}{A \code{numeric} matrix of observations with with horizontal axis being time, and with each column be the vectorized adjacency matrix.}

\item{data_mat2}{A \code{numeric} matrix of observations with with horizontal axis being time, and with each column be the vectorized adjacency matrix (data_mat1 and data_mat2 are independent and have the same dimensions ).}

\item{s}{A \code{integer} scalar of starting index.}

\item{e}{A \code{integer} scalar of ending index.}

\item{delta}{A positive \code{integer} scalar of minimum spacing.}

\item{level}{Should be fixed as 0.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item S:           A vector of estimated changepoints (sorted in strictly increasing order).
 \item Dval:        A vector of values of CUSUM statistic based on KS distance.
 \item Level:       A vector representing the levels at which each change point is detected.
 \item Parent:      A matrix with the starting indices on the first row and the ending indices on the second row.
}
}
\description{
Perform binary segmentation for network change points detection.
}
\examples{
y = c(rnorm(100, 0, 1), rnorm(100, 10, 10), rnorm(100, 40, 10))
}
\author{

}
