% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean.R
\name{BS.univar.CPD}
\alias{BS.univar.CPD}
\title{Univariate mean change point detection based on standard binary segmentation with tuning parameter selected by sSIC.}
\usage{
BS.univar.CPD(y, delta)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{delta}{A positive \code{integer} scalar of minimum spacing.}

\item{Alpha}{A \code{integer} vector of starting indices of random intervals.}

\item{Beta}{A \code{integer} vector of ending indices of random intervals.}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric} vector of estimated changepoint locations.
}
\description{
Perform Univariate mean change point detection based on standard binary segmentation. The threshold parameter tau for BS is automatically selected based on the sSIC score defined in Equation (4) in Fryzlewicz (2014).
}
\examples{
y = c(rnorm(100, 0, 1), rnorm(100, 2, 1), rnorm(100, 0, 1))
temp = BS.univar.CPD(y, delta = 5)
plot.ts(y)
points(x = temp, y = y[temp], col = "red")

}
\references{
Fryzlewicz (2014), Wild binary segmentation for multiple change-point detection,  Ann. Statist. 42(6): 2243-2281 (December 2014). DOI: 10.1214/14-AOS1245
}
\author{
Haotian Xu
}
