% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean.R
\name{BS.univar}
\alias{BS.univar}
\title{Standard binary segmentation for univariate mean change points detection}
\usage{
BS.univar(y, s, e, delta, level = 0, ...)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{s}{A \code{integer} scalar of starting index.}

\item{e}{A \code{integer} scalar of ending index.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{level}{Should be fixed as 0.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item S           A vector of estimated changepoints (sorted in strictly increasing order).
 \item Dval        A vector of values of CUSUM statistic based on KS distance.
 \item Level       A vector representing the levels at which each change point is detected.
 \item ...         Additional parameters.
}
}
\description{
TO DO
}
\examples{
y = c(rnorm(100, 0, 1), rnorm(100, 10, 10), rnorm(100, 40, 10))
temp = BS.univar(y, 1, 300, 5)
plot.ts(y)
points(x = tail(temp$S[order(temp$Dval)],20), y =rep(0, 20), col = "red")
}
\author{
Haotian Xu
}
