% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_polynomial.R
\name{coef.repara}
\alias{coef.repara}
\title{Coefficients reparametrization for polynomials with different bases}
\usage{
coef.repara(coef_vec, cpt1, cpt2, n)
}
\arguments{
\item{coef_vec}{A \code{numeric} vector of coefficients for polynomials associated with cpt1.}

\item{cpt1}{An \code{integer} scalar of the first change point.}

\item{cpt2}{An \code{integer} scalar of the second change point.}

\item{n}{An \code{integer} scalar of sample size.}
}
\value{
A vector of transformed coefficients for polynomials associated with cpt2.
}
\description{
Compute the transformed coefficients for polynomials with different bases (currently, only the linear, quadratic functions and cubic functions are considered).
}
\examples{
TO DO
}
\references{
Yu, Y., & Chatterjee, S. (2020). Localising change points in piecewise polynomials of general degrees. arXiv preprint arXiv:2007.09910.
}
\author{
Haotian Xu
}
