% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD_regression.R
\name{CV.DP.regression}
\alias{CV.DP.regression}
\title{Cross-Validation of Dynamic Programming algorithm for regression change points detection by l0 penalty}
\usage{
CV.DP.regression(gamma, delta, y, X, lambda, ...)
}
\arguments{
\item{gamma}{A \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{y}{A \code{numeric} vector of observations.}

\item{X}{A \code{numeric} matrix of covariates.}

\item{lambda}{A \code{numeric} scalar of tuning parameter for the lasso penalty.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
TO DO
}
\author{

}
