% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD_regression.R
\name{CV.search.DP.LR}
\alias{CV.search.DP.LR}
\title{Grid search based on Cross-Validation (all tuning parameters (gamma, lambda and zeta)) of local refinement for regression.}
\usage{
CV.search.DP.LR(y, X, gamma.set, lambda.set, zeta.set, delta, eps = 0.001)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{X}{A \code{numeric} matrix of covariates.}

\item{gamma.set}{A \code{numeric} vector of candidate tuning parameter associated with the l0 penalty.}

\item{lambda.set}{A \code{numeric} vector of candidate tuning parameter for the lasso penalty.}

\item{zeta.set}{A \code{numeric} vector of candidate tuning parameter for the group lasso.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item{cpt_hat}: A list of vector of estimated changepoints (sorted in strictly increasing order).
 \item{K_hat}: A list of scalar of number of estimated changepoints.
 \item{test_error}: A list of vector of testing errors (each column corresponding to each lambda, and each column corresponding to each gamma).
 \item{train_error}: A list of vector of training errors.
}
}
\description{
TO DO
}
\examples{
set.seed(123)
data = simu.change.regression(10, c(10, 30, 40, 70, 90), 30, 100, 1, 9)
temp = CV.search.DP.LR(data$y, data$X, gamma.set = c(1, 2, 5, 10, 20, 30), lambda.set = c(0.01, 0.05, 0.1, 0.5, 1, 2, 3, 4, 5), zeta.set = c(0.1, 1, 10), delta = 3)
temp[3,]
zeta_sel = 1
gamma_sel = 1
lambda_sel = 0.1
cpt_hat_init = part2local(DP.regression(data$y, X = data$X, gamma = gamma_sel, lambda = lambda_sel, delta = 3)$partition)
cpt_hat_LR = local.refine.regression(cpt_hat_init, data$y, X = data$X, zeta = zeta_sel, 1/3)
}
\author{
Daren Wang
}
