% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD_regression.R
\name{CV.search.DP.regression}
\alias{CV.search.DP.regression}
\title{Grid search based on Cross-Validation of Dynamic Programming for regression change points detection via l0 penalty}
\usage{
CV.search.DP.regression(y, X, gamma.set, lambda.set, delta)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{X}{A \code{numeric} matrix of covariates.}

\item{gamma.set}{A \code{numeric} vector of candidate tuning parameter associated with the l0 penalty.}

\item{lambda.set}{A \code{numeric} vector of candidate tuning parameter for the lasso penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item{cpt_hat}: A list of vector of estimated change points locations (sorted in strictly increasing order).
 \item{K_hat}: A list of scalar of number of estimated change points.
 \item{test_error}: A list of vector of testing errors.
 \item{train_error}: A list of vector of training errors.
}
}
\description{
TO DO
}
\examples{
data = simu.change.regression(10, c(10, 30, 40, 70, 90), 30, 100, 1, 9)
CV.search.DP.regression(data$y, data$X, gamma.set = 1:5, lambda.set = 1:5, delta = 5)
cpt.init = part2local(DP.regression(data$y, X = data$X, gamma = 4, lambda = 2, delta = 5)$partition)
local.refine.regression(cpt.init, data$y, X = data$X, 1, 1/3)
}
\author{
Daren Wang
}
