% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{D_P}
\alias{D_P}
\title{Dynamic programming for l0 change points detection (RCPP)}
\usage{
D_P(y, gamma, ...)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{gamma}{A \code{numeric} scalar corresponding to the tuning parameter associated with the l0 penalty.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
y = rnorm(300)+ c(rep(0,130),rep(-1,20),rep(1,20),rep(0,130))
D_P(y, 1)
}
\author{
Haotian Xu
}
