% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean.R
\name{D_P_univar}
\alias{D_P_univar}
\title{Dynamic programming for univariate mean change points detection by l0 penalty (RCPP)}
\usage{
D_P_univar(gamma, delta, y, ...)
}
\arguments{
\item{gamma}{A \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{y}{A \code{numeric} vector of observations.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
y = rnorm(300)+ c(rep(0,130),rep(-1,20),rep(1,20),rep(0,130))
D_P_univar(1, 5, y)
}
\author{
Haotian Xu
}
