% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data.generate.k.fluctuate2}
\alias{data.generate.k.fluctuate2}
\title{Simulate Model1 in [10] and follows Simulations 4.2}
\usage{
data.generate.k.fluctuate2(d0, cpt.true, p, n, sigma, kappa)
}
\arguments{
\item{d0}{A \code{numeric} scalar of number of nonzero coefficients.}

\item{p}{A \code{integer} scalar of dimensionality.}

\item{n}{A \code{integer} scalar of sample size.}

\item{sigma}{A \code{numeric} scalar of error standard deviation.}

\item{kappa}{A \code{numeric} scalar of minimum jump size in terms of the l2 norm.}

\item{cpt.ture}{A \code{integer} vector of true changepoints (sorted in strictly increasing order).}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item cpt.true    A vector of true changepoints (sorted in strictly increasing order).
 \item X           A p-by-n matrix of (random) design matrix.
 \item y           A n-dim vector of response variable.
 \item betafullmat A p-by-n matrix of coefficients.
 \item ...         Additional parameters.
}
}
\description{
TO DO
}
\examples{
TO DO
}
\author{

}
