% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{distanceR}
\alias{distanceR}
\title{Prediction error in l2 norm for the lasso [10]}
\usage{
distanceR(s, e, y, X, lambda, delta, ...)
}
\arguments{
\item{s}{A \code{integer} scalar of starting index.}

\item{e}{A \code{integer} scalar of ending index.}

\item{y}{A \code{numeric} vector of response variable.}

\item{X}{A \code{numeric} matrix of covariates.}

\item{lambda}{A \code{numeric} scalar of lasso penalty.}

\item{delta}{A \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric} scalar of prediction error in l2 norm.
}
\description{
TO DO
}
\examples{
TO DO
}
\author{

}
