% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DP}
\alias{DP}
\title{Dynamic programming for l0 change points detection}
\usage{
DP(y, gamma, ...)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{gamma}{A \code{numeric} scalar corresponding to the tuning parameter associated with the l0 penalty.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
y = c(rep(0, 100), rep(1, 100))
DP(y, 1)
}
\author{
Haotian Xu
}
