% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{DP.VAR1}
\alias{DP.VAR1}
\title{Dynamic programming for VAR1 change points detection by l0 penalty}
\usage{
DP.VAR1(DATA, gamma, delta, lambda, ...)
}
\arguments{
\item{DATA}{A \code{numeric} matrix of observations.}

\item{gamma}{A \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{lambda}{A \code{numeric} scalar of tuning parameter for lasso penalty.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
p = 20
sigma = 1
n = 100
A = matrix(rnorm(p*p), nrow = p)
DATA = simu.change.VAR1(sigma, p, n, A)
DP.VAR1(DATA, gamma = 1, delta = 5, lambda = 1)
}
\author{
Haotian Xu
}
