% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findcpts.R
\name{findcpts}
\alias{findcpts}
\title{Changpoint Finder}
\usage{
findcpts(
  x,
  dec = sqrt(2),
  minl = 2L,
  thr = 0,
  par = 40L,
  statsonly = FALSE,
  penalty = "BIC"
)
}
\arguments{
\item{x}{Data vector: vector of length n > 2}

\item{dec}{Decay rate: Rate of how fast the different layers will decrease in size. Default is set to sqrt(2).}

\item{minl}{Minimal theoretical interval-length: The minimal size of intervals to be considered. Default is set to 2.}

\item{thr}{Threshold: Only intervals with test statistic greater equal than this threshold will be considered. Should be greater equal to 0. Default is set to 0.}

\item{par}{Cutoff parameter: Cutoff between full search and optimistic search. Intervals with lower length than the cutoff will be evaulated by full search, the rest will be evaluated by optimistic search, which performs faster for large intervals. Default is set to 40.}

\item{statsonly}{Binary paramter: If TRUE the function will only calculate and output the test statics for each interval and omit the changepoint detection. Default is set to FALSE.}

\item{penalty}{Penalty Type: Two kind of penalty types are possible. If penalty="BIC" the Bayesian information criterion penalty type is used. If penalty="MBIC" a modified BIC criterion is used. Default is set to "BIC".}
}
\value{
A list with the following (except if statsonly=TRUE):
\item{Cpts}{Matrix with 6 columns: changepoint found, start of the location where the changpoint was found, end of the location where the changepoint was found, test statistic to the interval from where the changepoint was taken from, residual sum of squares including the new changepoint and the corresponding criterion penalty}
\item{Intervals}{Matrix with 4 columns: start point of the interval, end point of the interval, optimal changepoint for that interval (the optimal splitting point is between output and output+1) and corresponding teststatistic}
\item{RSS}{Total residual sum of squares}
\item{OptByCrit}{The optimal value for the criterion penalty. If the values fo OptByCrit and RSS coincide then no changepoint is relevant (will not show up if no changepoints are over the threshold)}
\item{OptCpts}{All changpoints up to the optimal value for the criterion penalty (will not show up if no changepoints are over the threshold)}
}
\description{
Function that finds changepoints in given data set x by evaluating a structural set of intervals which are a subset of the powerset from 1 to length(x)
}
\examples{
\donttest{
x=rnorm(10)
findcpts(x)
}
}
