% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{gen.cov.mat}
\alias{gen.cov.mat}
\title{Generate population covariance matrix with dimension p.}
\usage{
gen.cov.mat(p, sigma2, type)
}
\arguments{
\item{p}{A \code{integer} scalar of dimensionality.}

\item{sigma2}{A positive \code{numeric} scalar representing the variance of each entry.}

\item{type}{3 types of covariance matrix are considered. 1: Diagonal structure; 2: Equal correlation structure; 3: Power decay structure.}
}
\value{
A \code{numeric} p-by-p matrix.
}
\description{
Generate population covariance matrix with dimension p.
}
\examples{
TO DO
}
\author{

}
