% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{gen.lower.coordinate}
\alias{gen.lower.coordinate}
\title{Generate coordinates of lower triangular matrix of dimension p.}
\usage{
gen.lower.coordinate(p)
}
\arguments{
\item{p}{A \code{integer} scalar of dimensionality.}
}
\value{
A \code{integer} (p*(p-1)/2)-dim vector representing the indices of the lower triangular entries (indices correspond to the vectorization by stacking columns).
}
\description{
Generate coordinates of lower triangular matrix of dimension p.
}
\examples{
gen.lower.coordinate(5)
}
\author{
Oscar Hernan Madrid Padilla
}
