% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{gen.missing}
\alias{gen.missing}
\title{Function to generate a matrix with values 0 or 1, where 0 indicating the entry is missing}
\usage{
gen.missing(pi_mat, symm = TRUE)
}
\arguments{
\item{pi_mat}{A \code{numeric} pxp matrix, for each entry, the value representing the probability of missing.}

\item{symm}{A \code{logic} scalar indicating if the output matrix needs to be symmetric.}
}
\value{
A \code{numeric} p x p matrix.
}
\description{
Function to generate a matrix with values 0 or 1, where 0 indicating the entry is missing
}
\examples{
p = 5
pi_mat = matrix(0.9, p, p)
eta_mat = gen.missing(pi_mat, symm = TRUE)
}
\author{
Haotian Xu
}
