% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_polynomial.R
\name{gen.piece.poly}
\alias{gen.piece.poly}
\title{Generate univariate data from piecewise polynomials}
\usage{
gen.piece.poly(init_coef_vec, cpt_vec, kappa_mat, n, sigma)
}
\arguments{
\item{init_coef_vec}{A \code{numeric} vector of coefficients for the first segment.}

\item{cpt_vec}{An \code{integer} vector of change points.}

\item{kappa_mat}{A \code{numeric} matrix where the i-th column represents the jump sizes for coefficients associated with the i-th change point.}

\item{n}{An \code{integer} scalar of sample size.}

\item{sigma}{A \code{numeric} scalar of standard deviation of error terms.}
}
\value{
A vector of data generated from piecewise polynomials.
}
\description{
Generate univariate data from piecewise polynomials (currently, only the linear, quadratic functions and cubic functions are considered).
}
\examples{
TO DO
}
\references{
Yu, Y., & Chatterjee, S. (2020). Localising change points in piecewise polynomials of general degrees. arXiv preprint arXiv:2007.09910.
}
\author{
Haotian Xu
}
