% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{Hausdorff.dist}
\alias{Hausdorff.dist}
\title{Bidirectional Hausdorff distance.}
\usage{
Hausdorff.dist(vec1, vec2, ...)
}
\arguments{
\item{vec1}{A \code{integer} vector forms a subset of {1, 2, ..., n}.}

\item{vec2}{A \code{integer} vector forms a subset of {1, 2, ..., n}.}

\item{...}{Additional arguments.}
}
\value{
An integer scalar of bidirectional Hausdorff distance.
}
\description{
Compute the bidirectional Hausdorff distance between two sets.
}
\examples{
vec1 = sample.int(1000, size = 50)
vec2 = sample.int(2000, size = 100)
Hausdorff.dist(vec1, vec2)
}
\author{
Daren Wang
}
