% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_nonparametric_L2.R
\name{kde.epan.eval}
\alias{kde.epan.eval}
\title{Multivariate kernel density estimation based on Epanechnikov kernel.}
\usage{
kde.epan.eval(x, bw, eval.points)
}
\arguments{
\item{x}{A \code{numeric} matrix of observations with horizontal axis being dimension, and vertical axis being time.}

\item{eval.points}{A \code{numeric} matrix of evaluated data points with horizontal axis being dimension, and vertical axis being time..}

\item{h}{A \code{numeric} scalar of bandwidth parameter.}
}
\value{
A numeric vector of evaluated densities.
}
\description{
Perform multivariate kernel density estimation and evaluated the estimated densities at specified points.
}
\examples{
n = 100
p = 10
x = matrix(rnorm(n*p), nrow = n)
h = 2*(1/n)^{1/(4+p)} # bandwith
kde.epan.eval(x, h, x)
}
\author{
Haotian Xu
}
