% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{latent_dirichlet_allocation}
\alias{latent_dirichlet_allocation}
\title{Estimates LDA topic model.}
\usage{
latent_dirichlet_allocation(corpus, latent_vars, niters = 1500L, alpha = 1,
  beta = 1)
}
\arguments{
\item{corpus}{Matrix corresponding to the DTM.}

\item{latent_vars}{List containing the set of variables to be
updated as part of the estimation procedure.}

\item{niters}{Number of iterations for LDA to run.}

\item{alpha}{Prior for theta.}

\item{beta}{Prior for phi.}
}
\value{
latent_vars Updated version of List that was taken
        as input.
}
\description{
Estimates a LDA topic model using collapsed Gibbs
             sampling.
}
\author{
\packageMaintainer{changepointsHD}
}
