% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{latent_dirichlet_allocation_ll}
\alias{latent_dirichlet_allocation_ll}
\title{Estimates log-likelihood for LDA.}
\usage{
latent_dirichlet_allocation_ll(corpus, latent_vars)
}
\arguments{
\item{corpus}{Matrix corresponding to the DTM.}

\item{latent_vars}{List containing the set of variables to be
updated as part of the estimation procedure.}
}
\value{
Log-likelihood estimate.
}
\description{
Generates the log-likelihood for a corresponding set of
             latent variables.
}
\author{
\packageMaintainer{changepoints}
}
