% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{local.refine}
\alias{local.refine}
\title{Local refinement [10]}
\usage{
local.refine(cpt.init, y, X, zeta, w = 0.9)
}
\arguments{
\item{cpt.init}{A \code{integer} vector of initial changepoints estimation (sorted in strictly increasing order).}

\item{y}{A \code{numeric} vector of response variable.}

\item{X}{A \code{numeric} matrix of covariates.}

\item{zeta}{A \code{numeric} scalar of lasso penalty.}

\item{w}{A \code{numeric} scalar of weight for interpolation.}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric} scalar of prediction error in l2 norm.
}
\description{
TO DO
}
\examples{
TO DO
}
\author{

}
