% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{local.refine.VAR1}
\alias{local.refine.VAR1}
\title{Local refinement for VAR1 change points detection.}
\usage{
local.refine.VAR1(cpt_init, DATA, zeta)
}
\arguments{
\item{cpt_init}{A \code{integer} vector of initial change points estimation (sorted in strictly increasing order).}

\item{DATA}{A \code{numeric} matrix of observations with with horizontal axis being time, and vertical axis being dimensions.}

\item{zeta}{A \code{numeric} scalar of lasso penalty.}
}
\value{
An \code{integer} vector of locally refined change points estimation.
}
\description{
Perform local refinement for VAR1 change points detection.
}
\references{
Wang, Yu, Rinaldo and Willett (2019) <arxiv:1909.06359>.
}
\seealso{
\code{\link{local.refine.CV.VAR1}}.
}
\author{
Daren Wang & Haotian Xu
}
