% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_nonparametric.R
\name{MNWBS.tune}
\alias{MNWBS.tune}
\title{A function to compute change points based on the multivariate nonparametic method.}
\usage{
MNWBS.tune(Y, W, Alpha, Beta, h, delta)
}
\arguments{
\item{Y}{A \code{numeric} matrix of observations with with horizontal axis being time, and vertical axis being dimension.}

\item{W}{A copy of the matrix Y, it can be Y itself.}

\item{Alpha}{A \code{integer} vector of starting indices of random intervals.}

\item{Beta}{A \code{integer} vector of ending indices of random intervals.}

\item{h}{A \code{numeric} scalar of bandwith parameter.}
}
\value{
A vector of estimated change points
}
\description{
A function to compute change points based on the multivariate nonparametic method.
}
\examples{
n = 200
v = c(floor(n/3), 2*floor(n/3)) # location of change points
p = 10
Y = matrix(0, p, n) # matrix for data
mu0 = rep(0, p) # mean of the data
mu1 = rep(0, p)
mu1[1:floor(p/2)] = 2
Sigma0 = diag(p) #Covariance matrices of the data
Sigma1 = diag(p)*2
# Generate data
for(t in 1:n){
  if(t < v[1] || t > v[2]){
     Y[,t] = MASS::mvrnorm(n = 1, mu0, Sigma0)
  }
  if(t >= v[1] && t < v[2]){
     Y[,t] = MASS::mvrnorm(n = 1, mu1, Sigma1)
  }
}## close for generate data
M = 25
intervals = WBS.intervals(M = M, lower = 1, upper = ncol(Y)) #Random intervals
K_max = 30
h = 5*(K_max*log(n)/n)^{1/p} # bandwith
S = MNWBS.tune(Y, Y, intervals$Alpha, intervals$Beta, h, delta = 10)
S
}
\author{
Oscar Hernan Madrid Padilla
}
