% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrowcpts.R
\name{narrowcpts}
\alias{narrowcpts}
\title{Changepoint Finder with narrow interval prioritization}
\usage{
narrowcpts(x, dec = sqrt(2), minl = 2L, thr = 1e-06, par = 40L)
}
\arguments{
\item{x}{Data vector: vector of length n > 2}

\item{dec}{Decay rate: Rate of how fast the different layers will decrease in size. Default is set to sqrt(2).}

\item{minl}{Minimal theoretical interval-length: The minimal size of intervals to be considered. Default is set to 2.}

\item{thr}{Positive Threshold: Only intervals with teststatistic greater equal than this threshold will be considered. Default is set to 1e-4.}

\item{par}{Cutoff parameter: Cutoff between full search and optimistic search. Intervals with lower length than the cutoff will be evaulated by full search, the rest will be evaluated by optimistic search, which performs faster for large intervals. Default is set to 40.}
}
\value{
A list with the following:
\item{Cpts}{Sorted changepoints found}
\item{Intervals}{Matrix with 4 columns: start point of the interval, end point of the interval, optimal changepoint for that interval (the optimal splitting point is between output and output+1) and corresponding teststatistic}
}
\description{
Function that finds changepoints in given data set x over a specific threshold and prioritizing narrow intervals
}
\examples{
\donttest{
x=rnorm(10)
narrowcpts(x,thr=0.1)
}
}
