% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_nonparametric.R
\name{NBS.tune}
\alias{NBS.tune}
\title{Standard binary segmentation for univariate nonparametric change points detection with tuning parameter selection.}
\usage{
NBS.tune(Y, N, delta)
}
\arguments{
\item{Y}{A \code{numeric} matrix of observations with horizontal axis being time, and vertical axis being multiple observations on each time point.}

\item{N}{A \code{integer} vector representing number of multiple observations on each time point.}

\item{delta}{A positive \code{integer} scalar of minimum spacing.}

\item{len_tau}{A \code{integer} scalar of number of candidate values considered to tune the threshold.}

\item{...}{Additional arguments.}
}
\value{
A vector of estimated change points (sorted in strictly increasing order).
}
\description{
Perform standard binary segmentation with tuning parameter selection based on sample splitting.
}
\examples{
Y = t(as.matrix(c(rnorm(100, 0, 1), rnorm(100, 0, 10), rnorm(100, 0, 40))))
N = rep(1, 300)
temp = NBS.tune(Y, N, 5)
}
\author{
Oscar Hernan Madrid Padilla & Haotian Xu
}
