% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{obj.func.lr}
\alias{obj.func.lr}
\title{An objective function to select the best splitting location in the local refinement, see eq(4) in [10]}
\usage{
obj.func.lr(s, e, eta, y, X, zeta)
}
\arguments{
\item{s}{A \code{integer} scalar of starting index.}

\item{e}{A \code{integer} scalar of ending index.}

\item{y}{A \code{numeric} vector of response variable.}

\item{X}{A \code{numeric} matrix of covariates.}

\item{zeta}{A \code{numeric} scalar of tuning parameter for the group lasso.}
}
\description{
An objective function to select the best splitting location in the local refinement, see eq(4) in [10]
}
