% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean_online.R
\name{online.univar.multi}
\alias{online.univar.multi}
\title{Online change point detection via CUSUM 4 (multiple change points).}
\usage{
online.univar.multi(y_vec, tau_vec, ...)
}
\arguments{
\item{y_vec}{A \code{numeric} vector of observations.}

\item{tau_vec}{A \code{numeric} vector of thresholds at time t>= 2.}

\item{...}{Additional arguments.}
}
\value{
An \code{integer} scalar of estimated change point location.
}
\description{
TO DO
}
\examples{
TO DO
}
\author{
Haotian Xu
}
