% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean_online.R
\name{online.univar.one}
\alias{online.univar.one}
\title{Online change point detection via CUSUM (single change point).}
\usage{
online.univar.one(
  y_vec,
  b_vec = NULL,
  train_vec = NULL,
  alpha = 0.05,
  permu_num = 100,
  ...
)
}
\arguments{
\item{y_vec}{A \code{numeric} vector of observations.}

\item{b_vec}{A \code{numeric} vector of thresholds b_t with t >= 2.}

\item{train_vec}{A \code{numeric} vector of training data from a pre-change distribution, which is only needed to when b_t is NULL in order to calibrate b_t.}

\item{...}{Additional arguments.}
}
\value{
An \code{integer} scalar of estimated change point location.
}
\description{
Perform online change point detection via CUSUM (single change point).
}
\examples{
TO DO
}
\author{
Haotian Xu
}
