% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean_online.R
\name{online.univar.one3}
\alias{online.univar.one3}
\title{Online change point detection via CUSUM (single change point, type 3).}
\usage{
online.univar.one3(y_vec, tau_vec, ...)
}
\arguments{
\item{y_vec}{A \code{numeric} vector of observations.}

\item{tau_vec}{A \code{numeric} vector of thresholds at time t>= 1.}

\item{...}{Additional arguments.}
}
\value{
An \code{integer} scalar of estimated change point location.
}
\description{
Perform online change point detection via CUSUM (single change point, type 3).
}
\examples{
TO DO
}
\author{
Haotian Xu
}
