% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{part2local}
\alias{part2local}
\title{Partition to localization}
\usage{
part2local(x)
}
\arguments{
\item{x}{A \code{numeric} vector of observations.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
y = c(rep(0, 100), rep(1, 100))
part2local(D_P_univar(1, 4, y)$partition)
}
\author{
Haotian Xu
}
