% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{print.BS}
\alias{print.BS}
\title{Function to print a BS object}
\usage{
\method{print}{BS}(x, ...)
}
\arguments{
\item{x}{A \code{BS} object.}

\item{...}{Additional arguments.}
}
\description{
Function to print a BS object
}
\author{
Haotian Xu
}
