% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_gradient_ll}
\alias{prox_gradient_ll}
\title{estimates the log-likeihood for the corresponding
       theta and data set}
\usage{
prox_gradient_ll(data, theta_i, regularizer)
}
\arguments{
\item{data}{N x P matrix corresponding to the raw data}

\item{theta_i}{estimate for theta}

\item{regularizer}{regularizing constant, lambda}
}
\value{
log-likelihood
}
\description{
estimates the log-likeihood for the corresponding
             theta and data set
}
\author{
Leland Bybee \email{leland.bybee@gmail.com}
}
