% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_gradient_mapping}
\alias{prox_gradient_mapping}
\title{Performs the proximal-gradient mapping operation to
       estimate a regularized version of the inverse cov
       matrix}
\usage{
prox_gradient_mapping(data, theta_start, update_w, update_change, regularizer,
  max_iter, tol)
}
\arguments{
\item{data}{N x P matrix corresponding to the raw data}

\item{theta_start}{initial value for theta estimate}

\item{update_w}{step size for prox-gradient}

\item{update_change}{proportion of update_w to keep when
the algorithm fails to successfully estimate theta}

\item{regularizer}{regularizing constant, lambda}

\item{max_iter}{number of mapping iterations}

\item{tol}{tolerance at which the algorithm stops running}
}
\value{
theta estimate
}
\description{
Performs the proximal-gradient mapping operation to
             estimate a regularized version of the inverse cov
             matrix
}
\author{
Leland Bybee \email{leland.bybee@gmail.com}
}
