% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rank_one}
\alias{rank_one}
\title{method for estimating single-changepoint using special
       structure of the GGM framework}
\usage{
rank_one(data, theta_int, buff = 10L, regularizer = 1, tau = -1L,
  max_iter = 25L, update_w = 1, update_change = 0.9, mapping_iter = 1L,
  tol = 1e-05)
}
\arguments{
\item{data}{N x P matrix corresponding to the raw data}

\item{buff}{distance to maintain from edge of sample}

\item{regularizer}{regularizing constant, lambda}

\item{tau}{initial estimate for change-point}

\item{max_iter}{maximum number of rank-one updates to be
run}

\item{update_w}{step size for prox-gradient}

\item{update_change}{proportion of update_w to keep when
the algorithm fails to successfully estimate theta}

\item{mapping_iter}{number of mapping iterations}

\item{tol}{tolerance at which the algorithm stops running}

\item{theta_init}{initial value for theta estimate}
}
\value{
List containing the estimated change-point and
        theta values
}
\description{
This is a method for estimating a single-changepoint
             which takes advantage of the special structure
             of the Gaussian graphical model.  It cannot take
             arbitrary black-box models like simulated_annealing
             or brute_force.  However, it can still be run within
             binary segmentation.
}
\author{
Leland Bybee \email{leland.bybee@gmail.com}
}
