% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_optimal_partitioning}
\alias{rcpp_optimal_partitioning}
\title{Optimal Partitioning Method}
\usage{
rcpp_optimal_partitioning(x, beta)
}
\arguments{
\item{x}{A numeric vector representing the data to segment.}

\item{beta}{A double value representing the penalty term for adding a new segment.}
}
\value{
A list with (1) the changepoint elements (each last index of each segment in \code{changepoints}), (2) a vector `\code{nb} saving the number of non-pruned elements at each iteration, (3) a vector \code{lastIndexSet} containing the non-pruned indices at the end of the algo and (4) a vector \code{costQ} saving the optimal cost at each time step.
}
\description{
This function computes the optimal partitioning of a given vector x with a given penalty term beta. It finds the optimal changepoints that minimize the cost function using dynamic programming.
}
\details{
Optimal Partitioning Method using C++
}
\examples{
n <- 100
data <- rep(c(0, 5, 2.5, 7), each = n) + rnorm(4 * n)
OP_Rcpp <- optimal_partitioning(data, 2 * log(length(data)))
plot_segmentation(data, OP_Rcpp$changepoints, title = "OP Segmentation")

}
