% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_smallest_valid_partitioning_VR}
\alias{rcpp_smallest_valid_partitioning_VR}
\title{SVP Method}
\usage{
rcpp_smallest_valid_partitioning_VR(y, gamma, test, pruning = TRUE)
}
\arguments{
\item{y}{A numeric vector representing the data to segment.}

\item{gamma}{A numeric value used as a threshold in the validation function and as a penalty for each segment.}

\item{test}{A function of the form `function(y, gamma)` returning TRUE if the segment is valid.}

\item{pruning}{Logical indicating whether to apply pruning to the candidate set during the dynamic programming. Default is TRUE.}
}
\value{
A list with (1) the changepoint elements (each last index of each segment in \code{changepoints}), (2) a vector `\code{nb} saving the number of non-pruned elements at each iteration, (3) a vector \code{lastIndexSet} containing the non-pruned indices at the end of the algo and (4) a vector \code{costQ} saving the optimal cost at each time step.
}
\description{
This function implements the SVP (Smallest Valid Partitioning) method in Rcpp.
}
\details{
SVP Method using C++
}
