% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD_regression.R
\name{simu.change.regression}
\alias{simu.change.regression}
\title{Simulate sparse regression model with changepoints in coefficients.}
\usage{
simu.change.regression(d0, cpt.true, p, n, sigma, kappa, ...)
}
\arguments{
\item{d0}{A \code{numeric} scalar of number of nonzero coefficients.}

\item{p}{An \code{integer} scalar of dimensionality.}

\item{n}{An \code{integer} scalar of sample size.}

\item{sigma}{A \code{numeric} scalar of error standard deviation.}

\item{kappa}{A \code{numeric} scalar of minimum jump size in terms of the l2 norm.}

\item{...}{Additional arguments.}

\item{cpt.ture}{An \code{integer} vector of true changepoints (sorted in strictly increasing order).}
}
\value{
A \code{list} with the structure:
\itemize{
 \item cpt.true    A vector of true changepoints (sorted in strictly increasing order).
 \item X           A p-by-n matrix of (random) design matrix.
 \item y           A n-dim vector of response variable.
 \item betafullmat A p-by-n matrix of coefficients.
 \item ...         Additional parameters.
}
}
\description{
Simulate sparse regression model with changepoints in coefficients under the setting of Simulations 4.2 [10].
}
\examples{
data = simu.change.regression(10, c(10, 30, 40, 70, 90), 30, 100, 1, 9)
}
\author{
Daren Wang
}
