% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{simu.change.VAR1}
\alias{simu.change.VAR1}
\title{Simulate a VAR1 model (without change point)}
\usage{
simu.change.VAR1(sigma, p, n, A, vzero = NULL, ...)
}
\arguments{
\item{sigma}{A \code{numeric} scalar representing error standard deviation.}

\item{p}{A \code{integer} scalar representing dimensionality.}

\item{n}{A \code{integer} scalar representing sample size.}

\item{A}{A \code{numeric} p-by-p matrix representing transition matrix of VAR1 model.}

\item{vzero}{A \code{numeric} vector representing the observation at time 0.}

\item{...}{Additional arguments.}
}
\value{
A p-by-n matrix.
}
\description{
TO DO
}
\examples{
p = 20
sigma = 1
n = 100
A = matrix(rnorm(p*p), nrow = p)
simu.change.VAR1(sigma, p, n, A)
}
\author{

}
