% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEPP.R
\name{simu.SEPP}
\alias{simu.SEPP}
\title{Simulate a (stable) SEPP model (without change point)}
\usage{
simu.SEPP(intercept, M, A, threshold, n, vzero = NULL, ...)
}
\arguments{
\item{intercept}{A \code{numeric} scalar representing the intercept of the model.}

\item{M}{A \code{integer} scalar representing dimensionality.}

\item{A}{A \code{numeric} M-by_M matrix representing the coefficient matrix.}

\item{threshold}{A \code{numeric} scalar representing the upper bound for each coordinate of X_t (for stability).}

\item{n}{A \code{integer} scalar representing sample size.}

\item{vzero}{A \code{numeric} vector representing the observation at time 0.}

\item{...}{Additional arguments.}
}
\value{
A p-by-n matrix.
}
\description{
TO DO
}
\examples{
p = 20
sigma = 1
n = 100
A = matrix(rnorm(p*p), nrow = p)
simu.change.VAR1(sigma, p, n, A)
}
\author{

}
