% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoints.R
\name{simulated_annealing}
\alias{simulated_annealing}
\title{Estimates a single change-point using the simulated annealing
       method}
\usage{
simulated_annealing(data, bbmod_init_vals, bbmod_method, bbmod_ll,
  niter = 500, min_beta = 1e-04, buff = 100,
  bbmod_method_params = list(), bbmod_ll_params = list())
}
\arguments{
\item{data}{matrix of actual data.}

\item{bbmod_init_vals}{initial estimates for arbitrary bbmod values}

\item{bbmod_method}{function corresponding to black box model itself}

\item{bbmod_ll}{function returning ll or comparable cost function for
black box model}

\item{niter}{number of simulated annealing iterations}

\item{min_beta}{lowest temperature}

\item{buff}{distance from edge of sample to be maintained}

\item{bbmod_method_params}{list of additional parameters for bbmod_method}

\item{bbmod_ll_params}{list of additional parameters for bbmod_ll}
}
\value{
List containing estimated change-point and thetas
}
\description{
Estimates a single change-point using the simulated annealing
             method
}
\author{
Leland Bybee \email{leland.bybee@gmail.com}
}
