% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{threshold.BS}
\alias{threshold.BS}
\title{Thresholding a BS object with threshold value tau.}
\usage{
threshold.BS(BS_object, tau, ...)
}
\arguments{
\item{BS_object}{A \code{BS} object.}

\item{tau}{A positive \code{numeric} scalar of thresholding value.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the following structure:
 \item{BS_tree}{A list of data.frame containing "current" indices, "parent" indices, "location" indices and "value" of CUSUM at all steps}
 \item{BS_tree_trimmed}{BS_tree with change points which do not satisfy the thresholding criteria removed}
 \item{cpt_hat}{A matrix contains change point locations, values of corresponding statistic, and levels at which each change point is detected}
}
\description{
Given a BS object, perform thresholding to find the change point locations.
}
\examples{
y = c(rnorm(100, 0, 1), rnorm(100, 10, 10), rnorm(100, 40, 10))
temp = BS.univar(y, 1, 300, 5)
plot.ts(y)
points(x = tail(temp$S[order(temp$Dval)],4), y = y[tail(temp$S[order(temp$Dval)],4)], col = "red")
threshold.BS(temp, 20)
}
\seealso{
\code{BS.univar}, \code{BS.cov}, \code{WBS.univar}, \code{WBS.network}
}
\author{
Haotian Xu
}
