% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD_regression.R
\name{trim_interval}
\alias{trim_interval}
\title{Interval trimming based on initial change point localisation.}
\usage{
trim_interval(n, cpt_init, w = 0.9)
}
\arguments{
\item{n}{An \code{integer} scalar corresponding to the sample size.}

\item{cpt_init}{An \code{integer} vector of initial changepoints estimation (sorted in strictly increasing order).}

\item{w}{A \code{numeric} scalar in (0,1) representing the weight for interval truncation.}
}
\value{
A matrix with each row be a trimmed interval.
}
\description{
Performing the interval trimming for local refinement.
}
\references{
Xu, Wang, Zhao and Yu (2022) <arXiv:2207.12453>.
}
\author{
Haotian Xu
}
