% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVP_ALGO_VALIDITYtests.R
\name{valid_Range}
\alias{valid_Range}
\title{Validity Test Based on Range}
\usage{
valid_Range(y, gamma)
}
\arguments{
\item{y}{A numeric vector representing a segment of the signal.}

\item{gamma}{A numeric threshold for the maximum allowed range.}
}
\value{
TRUE if the range is less than or equal to gamma.
}
\description{
Checks whether the range (max - min) of the segment is below a threshold.
}
\details{
Validity Test Based on Range
}
