% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_tests.R
\name{valid_RANGE_SLACK}
\alias{valid_RANGE_SLACK}
\title{Validity Test Based on Trimmed Range}
\usage{
valid_RANGE_SLACK(y, gamma)
}
\arguments{
\item{y}{A numeric vector representing a segment of the signal.}

\item{gamma}{A numeric threshold for the trimmed range.}
}
\value{
TRUE if the trimmed range is less than or equal to gamma, or if the segment is too short.
}
\description{
Applies a slack range test using trimmed minimum and maximum (ignores smallest and largest 3 values).
}
\details{
Validity Test Based on Trimmed Range
}
