% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity_tests.R
\name{valid_VAR}
\alias{valid_VAR}
\title{Validity Test Based on Variance}
\usage{
valid_VAR(y, gamma)
}
\arguments{
\item{y}{A numeric vector representing a segment of the signal.}

\item{gamma}{A numeric threshold for the maximum allowed variance.}
}
\value{
TRUE if the variance is less than or equal to gamma, or if the segment is very short.
}
\description{
Checks if the variance of a segment is lower than or equal to a threshold.
}
\details{
Validity Test Based on Variance
}
