% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{WBS.intervals}
\alias{WBS.intervals}
\title{Generate random intervals for WBS.}
\usage{
WBS.intervals(M, lower = 1, upper, ...)
}
\arguments{
\item{M}{A positive \code{integer} scalar of number of random intervals.}

\item{lower}{A positive \code{integer} scalar of lower bound of random intervals.}

\item{upper}{A positive \code{integer} scalar of upper bound of random intervals.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the following structure:
 \item{Alpha}{A M-dim vector representing the starting indices of random intervals}
 \item{Beta}{A M-dim vector representing the ending indices of random intervals}
}
\description{
Generate random intervals for WBS.
}
\examples{
WBS.intervals(120, lower = 1, upper = 300)
}
\author{
Oscar Hernan Madrid Padilla
}
