% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean.R
\name{WBS.univar.CPD}
\alias{WBS.univar.CPD}
\title{Univariate mean change point detection based on wild binary segmentation with tuning parameter selected by sSIC.}
\usage{
WBS.univar.CPD(y, Alpha, Beta, delta)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{Alpha}{A \code{integer} vector of starting indices of random intervals.}

\item{Beta}{A \code{integer} vector of ending indices of random intervals.}

\item{delta}{A positive \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric} vector of estimated changepoint locations.
}
\description{
Perform Univariate mean change point detection based on wild binary segmentation. The threshold parameter tau for WBS is automatically selected based on the sSIC score defined in Equation (4) in Fryzlewicz (2014).
}
\examples{
y = c(rnorm(100, 0, 1), rnorm(100, 1, 1), rnorm(100, 0, 1))
intervals = WBS.intervals(M = 500, lower = 1, upper = length(y))
temp = WBS.univar.CPD(y, intervals$Alpha, intervals$Beta, delta = 5)
plot.ts(y)
points(x = temp, y = y[temp], col = "red")

}
\references{
Fryzlewicz (2014), Wild binary segmentation for multiple change-point detection,  Ann. Statist. 42(6): 2243-2281 (December 2014). DOI: 10.1214/14-AOS1245
}
\author{
Haotian Xu
}
