% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{X.glasso.converter}
\alias{X.glasso.converter}
\title{Convert a p-by-n design submatrix X with partial consecutive observations into a n-by-(2p) matrix, which fits the group lasso, see eq(4) in [10]}
\usage{
X.glasso.converter(X, eta, s_ceil)
}
\arguments{
\item{X}{A \code{numeric} matrix of covariates with partial consecutive observations.}

\item{eta}{A \code{integer} scalar of splitting index.}

\item{s_ceil}{A \code{numeric} matrix of covariates.}
}
\value{
A n-by-(2p) matrix
}
\description{
Convert a p-by-n design submatrix X with partial consecutive observations into a n-by-(2p) matrix, which fits the group lasso, see eq(4) in [10]
}
