% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean.R
\name{BS.univar}
\alias{BS.univar}
\title{Standard binary segmentation for univariate mean change points detection.}
\usage{
BS.univar(y, s, e, delta = 2, level = 0, ...)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{s}{A \code{integer} scalar of starting index.}

\item{e}{A \code{integer} scalar of ending index.}

\item{delta}{A positive \code{numeric} scalar of minimum spacing.}

\item{level}{Should be fixed as 0.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item{S}: A vector of estimated change point locations (sorted in strictly increasing order).
 \item{Dval}: A vector of values of CUSUM statistic based on KS distance.
 \item{Level}: A vector representing the levels at which each change point is detected.
 \item{Parent}: A matrix with the starting indices on the first row and the ending indices on the second row.
}
}
\description{
Perform standard binary segmentation for univariate mean change points detection.
}
\examples{
y = c(rnorm(100, 0, 1), rnorm(100, 1, 1), rnorm(100, 0, 1))
temp = BS.univar(y, 1, length(y), delta = 10)
plot.ts(y)
points(x = tail(temp$S[order(temp$Dval)],4),
       y = y[tail(temp$S[order(temp$Dval)],4)], col = "red")
threshold.BS(temp, 2)
}
\author{
Haotian Xu
}
