\name{changepointsHD-package}
\alias{changepointsHD-package}
\alias{changepointsHD}
\docType{package}
\title{
\packageTitle{changepointsHD}
}
\description{
\packageDescription{changepointsHD}
}
%\details{
%\tabular{ll}{
%Package: \tab \packagePackage{changepoints} \cr
%Type: \tab \packageType{changepoints} \cr
%Version: \tab \packageVersion{changepoints} \cr
%Date: \tab \packageDate{changepoints} \cr
%Licence: \tab packageLicense{changepoints} \cr
%}
\author{
\packageAuthor{changepointsHD}

Maintainer: \packageMaintainer{changepointsHD}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
%~~ in the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
