% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean.R
\name{CV.DP.univar}
\alias{CV.DP.univar}
\title{Cross-Validation of Dynamic Programming algorithm for univariate mean change points detection by l0 penalty}
\usage{
CV.DP.univar(y, gamma, delta, ...)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{gamma}{A \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{delta}{A positive \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
TO DO
}
\author{

}
