% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{CV.DP.VAR1}
\alias{CV.DP.VAR1}
\title{Cross-Validation of Dynamic Programming algorithm for VAR1 change points detection via l0 penalty}
\usage{
CV.DP.VAR1(DATA, gamma, lambda, delta, ...)
}
\arguments{
\item{DATA}{A \code{numeric} matrix of observations with with horizontal axis being time, and vertical axis being dimensions.}

\item{gamma}{A positive \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{lambda}{A positive \code{numeric} scalar of tuning parameter for the lasso penalty.}

\item{delta}{A positive \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
TO DO
}
\author{

}
