% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD_regression.R
\name{CV.search.DP.regression}
\alias{CV.search.DP.regression}
\title{Grid search based on Cross-Validation of Dynamic Programming for regression change points detection via l0 penalty}
\usage{
CV.search.DP.regression(y, X, gamma.set, lambda.set, delta, eps = 0.001)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{X}{A \code{numeric} matrix of covariates.}

\item{gamma.set}{A \code{numeric} vector of candidate tuning parameter associated with the l0 penalty.}

\item{lambda.set}{A \code{numeric} vector of candidate tuning parameter for the lasso penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item{cpt_hat}: A list of vector of estimated changepoints (sorted in strictly increasing order).
 \item{K_hat}: A list of scalar of number of estimated changepoints.
 \item{test_error}: A list of vector of testing errors (each column corresponding to each lambda, and each column corresponding to each gamma).
 \item{train_error}: A list of vector of training errors.
}
}
\description{
TO DO
}
\examples{
set.seed(123)
data = simu.change.regression(10, c(10, 30, 40, 70, 90), 30, 100, 1, 9)
temp = CV.search.DP.regression(data$y, data$X, gamma.set = c(1, 2, 5, 10, 20, 30), lambda.set = c(0.01, 0.05, 0.1, 0.5, 1, 2, 3, 4, 5), delta = 2)
temp$test_error # test error result
min_idx = as.vector(arrayInd(which.min(result$test_error), dim(result$test_error))) # find the indices of gamma.set and lambda.set which minimizes the test error
cpt.init = unlist(temp$cpt_hat[min_idx[1], min_idx[2]])
local.refine.regression(cpt.init, data$y, X = data$X, 1, 1/3)
}
\author{
Daren Wang
}
