% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD_regression.R
\name{CV.search.DP.regression}
\alias{CV.search.DP.regression}
\title{Perform grid search based on Cross-Validation of Dynamic Programming algorithm for regression change points detection by l0 penalty}
\usage{
CV.search.DP.regression(lambda.set, gamma.set, y, X, delta)
}
\arguments{
\item{lambda.set}{A \code{numeric} vector of candidate tuning parameter for the lasso penalty.}

\item{gamma.set}{A \code{numeric} vector of candidate tuning parameter associated with the l0 penalty.}

\item{y}{A \code{numeric} vector of observations.}

\item{X}{A \code{numeric} matrix of covariates.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
TO DO
}
\author{

}
