% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean.R
\name{CV.search.DP.univar}
\alias{CV.search.DP.univar}
\title{Grid search for dynamic programming to select the tuning parameter through Cross-Validation.}
\usage{
CV.search.DP.univar(y, gamma.set, delta, ...)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{gamma.set}{A \code{numeric} vector of candidate tuning parameter associated with the l0 penalty.}

\item{delta}{A positive \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item{cpt_hat}: A list of vector of estimated change points locations (sorted in strictly increasing order).
 \item{K_hat}: A list of scalar of number of estimated change points.
 \item{test_error}: A list of vector of testing errors.
 \item{train_error}: A list of vector of training errors.
}
}
\description{
Perform grid search for dynamic programming to select the tuning parameter through Cross-Validation.
}
\examples{
y = rnorm(300) + c(rep(0,130),rep(1,20),rep(0,20),rep(1,130))
gamma.set = 3:9
DP_result = CV.search.DP.univar(y, gamma.set, delta = 5)
min_idx = which.min(DP_result$test_error)
part2local(DP.univar(y, gamma.set[min_idx], delta = 5)$partition)
}
\author{
Daren Wang and Haotian Xu
}
