% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean.R
\name{CV.search.DP.univar}
\alias{CV.search.DP.univar}
\title{Perform grid search based on Cross-Validation of Dynamic Programming algorithm for univariate mean change points detection by l0 penalty}
\usage{
CV.search.DP.univar(gamma.set, delta, y, ...)
}
\arguments{
\item{gamma.set}{A \code{numeric} vector of candidate tuning parameter associated with the l0 penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{y}{A \code{numeric} vector of observations.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
TO DO
}
\author{

}
