% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{CV.search.DP.VAR1}
\alias{CV.search.DP.VAR1}
\title{Grid search based on Cross-Validation of Dynamic Programming for regression change points detection via l0 penalty}
\usage{
CV.search.DP.VAR1(DATA, gamma.set, lambda.set, delta, ...)
}
\arguments{
\item{DATA}{A \code{numeric} matrix of observations with with horizontal axis being time, and vertical axis being dimensions.}

\item{gamma.set}{A \code{numeric} vector of candidate tuning parameter associated with the l0 penalty.}

\item{lambda.set}{A \code{numeric} vector of candidate tuning parameter for the lasso penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
TO DO
}
\author{

}
