% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{CV.search.DP.VAR1}
\alias{CV.search.DP.VAR1}
\title{Grid search based on Cross-Validation of Dynamic Programming for regression change points detection via l0 penalty}
\usage{
CV.search.DP.VAR1(DATA, gamma.set, lambda.set, delta, eps = 0.001, ...)
}
\arguments{
\item{DATA}{A \code{numeric} matrix of observations with with horizontal axis being time, and vertical axis being dimensions.}

\item{gamma.set}{A \code{numeric} vector of candidate tuning parameter associated with the l0 penalty.}

\item{lambda.set}{A \code{numeric} vector of candidate tuning parameter for the lasso penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
Row: lambda.set; column: gamma.set
}
\description{
TO DO
}
\examples{
p = 20
sigma = 1
n = 20
v1 = 2*(seq(1,p,1)\%\%2) - 1
v2 = -v1
AA = matrix(0, nrow = p, ncol = p-2)
A1 = cbind(v1,v2,AA)*0.1
A2 = cbind(v2,v1,AA)*0.1
A3 = A1
data = simu.VAR1(sigma, p, 2*n+1, A1)
data = cbind(data, simu.VAR1(sigma, p, 2*n, A2, vzero=c(data[,ncol(data)])))
data = cbind(data, simu.VAR1(sigma, p, 2*n, A3, vzero=c(data[,ncol(data)])))
gamma.set = c(0.1, 1)
lambda.set = c(0.1, 1, 2, 3)
temp = CV.search.DP.VAR1(data, gamma.set, lambda.set, delta = 10)
temp$test_error # test error result
min_idx = as.vector(arrayInd(which.min(temp$test_error), dim(temp$test_error))) # find the indices of gamma.set and lambda.set which minimizes the test error
cpt.init = unlist(temp$cpt_hat[min_idx[1], min_idx[2]])
}
\author{

}
