% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_polynomial.R
\name{DP.poly}
\alias{DP.poly}
\title{Dynamic programming algorithm for univariate polynomials change points detection.}
\usage{
DP.poly(y, r, gamma, delta, ...)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{r}{An \code{integer} scalar order of polynomials.}

\item{gamma}{A \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item partition   A vector of the best partition.
 \item yhat        A vector of mean estimation for corresponding to the best partition.
}
}
\description{
Perform dynamic programming algorithm for univariate polynomials change points detection.
}
\examples{
data = simu.change.regression(10, c(10, 30, 40, 70, 90), 30, 100, 1, 9)
DP.poly(2, 5, data$y, X = data$X, lambda = 1)
}
\author{

}
