% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DP}
\alias{DP}
\title{Dynamic programming for l0 change points detection}
\usage{
DP(gamma, delta, y, X = NULL, lambda = NULL, ...)
}
\arguments{
\item{gamma}{A \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{delta}{A \code{numeric} scalar of minimum spacing.}

\item{y}{A \code{numeric} vector of observations.}

\item{X}{A \code{numeric} matrix of covariates. If missing, then the mean change of y is considered.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
y = c(rep(0, 100), rep(1, 100))
DP(y, 1)
}
\author{
Haotian Xu
}
