% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD_regression.R
\name{DP.regression}
\alias{DP.regression}
\title{Dynamic programming for regression change points detection by l0 penalty}
\usage{
DP.regression(gamma, delta, y, X, lambda, ...)
}
\arguments{
\item{gamma}{A \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{y}{A \code{numeric} vector of observations.}

\item{X}{A \code{numeric} matrix of covariates. If missing, then the mean change of y is considered.}

\item{lambda}{A \code{numeric} scalar of tuning parameter for the lasso penalty.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
data = simu.change.regression(10, c(10, 30, 40, 70, 90), 30, 100, 1, 9)
DP.regression(2, 5, data$y, X = data$X, lambda = 1)
}
\author{
Haotian Xu
}
