% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD_regression.R
\name{DP.regression}
\alias{DP.regression}
\title{Dynamic programming algorithm for regression change points detection through l0 penalty.}
\usage{
DP.regression(y, X, gamma, lambda, delta, eps = 0.001, ...)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{X}{A \code{numeric} matrix of covariates.}

\item{gamma}{A positive \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{lambda}{A positive \code{numeric} scalar of tuning parameter for the lasso penalty.}

\item{delta}{A positive \code{integer} scalar of minimum spacing.}

\item{eps}{A \code{numeric} scalar of precision level for convergence of lasso.}

\item{...}{Additional arguments.}
}
\value{
A vector of the best partition.
}
\description{
Perform dynamic programming algorithm for regression change points detection through l0 penalty.
}
\examples{
data = simu.change.regression(10, c(10, 30, 40, 70, 90), 30, 100, 1, 9)
temp = DP.regression(data$y, X = data$X, gamma = 2, lambda = 1, delta = 5)
cpt_hat = part2local(temp$partition)
}
\references{
Rinaldo, A., Wang, D., Wen, Q., Willett, R., & Yu, Y. (2021, March). Localizing changes in high-dimensional regression models. In International Conference on Artificial Intelligence and Statistics (pp. 2089-2097). PMLR.
}
\author{
Daren Wang & Haotian Xu
}
