% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEPP.R
\name{DP.SEPP}
\alias{DP.SEPP}
\title{Dynamic programming for VAR1 change points detection through l0 penalty.}
\usage{
DP.SEPP(gamma, delta, delta2, DATA, intercept, lambda, threshold, ...)
}
\arguments{
\item{gamma}{A \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{DATA}{A \code{numeric} matrix of observations.}

\item{lambda}{A \code{numeric} scalar of tuning parameter for lasso penalty.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
Perform dynamic programming for VAR1 change points detection through l0 penalty.
}
\examples{
parti = DP.SEPP(gamma = 1, delta = 5, delta2 = 600, DATA, intercept = 1/2, lambda = 100, threshold = 6)$partition
localization = part2local(parti)
}
\author{
Haotian Xu
}
