% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DP.univar}
\alias{DP.univar}
\title{Dynamic programming for univariate mean change points detection by l0 penalty}
\usage{
DP.univar(gamma, delta, y, ...)
}
\arguments{
\item{gamma}{A \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{y}{A \code{numeric} vector of observations.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
y = c(rep(0, 100), rep(1, 100))
DP.univar(1, 4, y)
}
\author{
Haotian Xu
}
