% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean.R
\name{DP.univar}
\alias{DP.univar}
\title{Dynamic programming for univariate mean change points detection.}
\usage{
DP.univar(y, gamma, delta, ...)
}
\arguments{
\item{y}{A \code{numeric} vector of observations.}

\item{gamma}{A \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{delta}{A positive \code{integer} scalar of minimum spacing.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item{partition}: {A vector of the best partition.}
 \item{yhat}: {A vector of mean estimation for corresponding to the best partition.}
}
}
\description{
Perform dynamic programming for univariate mean change points detection through l0 penalty
}
\examples{
y = rnorm(300) + c(rep(0,130),rep(-1,20),rep(1,20),rep(0,130))
DP.univar(y, 1, 5)
}
\author{
Haotian Xu
}
