% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{DP.VAR1}
\alias{DP.VAR1}
\title{Dynamic programming for VAR1 change points detection by l0 penalty}
\usage{
DP.VAR1(DATA, gamma, delta, lambda, ...)
}
\arguments{
\item{DATA}{A \code{numeric} matrix of observations.}

\item{gamma}{A \code{numeric} scalar of the tuning parameter associated with the l0 penalty.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{lambda}{A \code{numeric} scalar of tuning parameter for lasso penalty.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
p = 20
sigma = 1
n = 30
v1 = 2*(seq(1,p,1)\%\%2) - 1
v2 = -v1
AA = matrix(0, nrow = p, ncol = p-2)
A1=cbind(v1,v2,AA)
A2=cbind(v2,v1,AA)
A3=A1
data = simu.change.VAR1(sigma, p, 2*n+1, A1)
data = cbind(simu.change.VAR1(sigma, p, 2*n, A2, vzero=c(data[,ncol(data)])))
data = cbind(simu.change.VAR1(sigma, p, 2*n, A3, vzero=c(data[,ncol(data)])))
parti = DP.VAR1(data, gamma = 1, delta = 5, lambda = 1)$partition
localization = part2local(parti)
}
\author{
Haotian Xu
}
