% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{gen.cov.mat}
\alias{gen.cov.mat}
\title{Generate population covariance matrix with dimension p.}
\usage{
gen.cov.mat(p, sigma2, type)
}
\arguments{
\item{p}{A \code{integer} scalar of dimensionality.}

\item{sigma2}{A positive \code{numeric} scalar representing the variance of each entry.}

\item{type}{Specify the type of a covariance matrix: Diagonal structure ("diagonal"); Equal correlation structure ("equal"); Power decay structure ("power").}
}
\value{
A \code{numeric} p-by-p matrix.
}
\description{
Generate population covariance matrix with dimension p.
}
\examples{
gen.cov.mat(p = 5, sigma2 = 1, type = "diagonal")
}
\author{
Haotian Xu
}
