% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{latent_dirichlet_allocation_ll}
\alias{latent_dirichlet_allocation_ll}
\title{Generates the log-likelihood for a corresponding set of
       latent variables}
\usage{
latent_dirichlet_allocation_ll(corpus, latent_vars)
}
\arguments{
\item{corpus}{matrix corresponding to the DTM}

\item{latent_vars}{List containing the set of variables to be
updated as part of the estimation procedure}
}
\value{
log-likelihood generated from latent vars
}
\description{
Generates the log-likelihood for a corresponding set of
             latent variables
}
\author{
Leland Bybee \email{leland.bybee@gmail.com}
}
