% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{local.refine.network}
\alias{local.refine.network}
\title{Local refinement for network change points detection [7]}
\usage{
local.refine.network(cpt_init, DATA, tau1, tau2 = Inf, w = 1/3, ...)
}
\arguments{
\item{cpt_init}{A \code{integer} vector of initial changepoints estimation (sorted in strictly increasing order).}

\item{DATA}{A \code{numeric} matrix of observations.}

\item{tau1}{A \code{numeric} scalar corresponding to the first parameter of the USVT.}

\item{tau2}{A \code{numeric} scalar corresponding to the second parameter of the USVT.}

\item{w}{A \code{numeric} scalar of weight for interpolation of starting and ending indices.}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric} vector of locally refined changepoint locations.
}
\description{
TO DO
}
\examples{
TO DO
}
\author{

}
