% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_mean.R
\name{local.refine.univar}
\alias{local.refine.univar}
\title{Local refinement for univariate mean changepoint detection.}
\usage{
local.refine.univar(cpt.init, y, w = 1/3)
}
\arguments{
\item{cpt.init}{An \code{integer} vector of initial changepoints estimation (sorted in strictly increasing order).}

\item{y}{A \code{numeric} vector of univariate time series.}

\item{w}{A \code{numeric} scalar of weight for interpolation.}

\item{zeta}{A \code{numeric} scalar of lasso penalty.}

\item{...}{Additional arguments.}
}
\value{
An \code{integer} vector of locally refined changepoint estimation.
}
\description{
Perform local refinement for univariate mean changepoint detection.
}
\examples{
TO DO
}
\author{

}
