% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{local.refine.VAR1}
\alias{local.refine.VAR1}
\title{Local refinement for VAR1 change points detection [11]}
\usage{
local.refine.VAR1(cpt.init, DATA, zeta.group, w = 1/3, ...)
}
\arguments{
\item{cpt.init}{A \code{integer} vector of initial changepoints estimation (sorted in strictly increasing order).}

\item{DATA}{A \code{numeric} matrix of observations with with horizontal axis being time, and vertical axis being dimensions.}

\item{zeta.group}{A \code{numeric} scalar of lasso penalty.}

\item{w}{A \code{numeric} scalar of weight for interpolation.}

\item{...}{Additional arguments.}
}
\value{
A \code{numeric} scalar of prediction error in l2 norm.
}
\description{
TO DO
}
\examples{
#TODO data = simu.change.regression(10, c(10, 30, 40, 70, 90), 30, 100, 1, 9)
cpt.init = part2local(DP.regression(2, 5, data$y, X = data$X, lambda = 2)$partition)$cpt
local.refine.regression(cpt.init, data$y, X = data$X, 1, 1/3)
}
\author{

}
