% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR.R
\name{local.refine.VAR1}
\alias{local.refine.VAR1}
\title{Local refinement for VAR1 change points detection.}
\usage{
local.refine.VAR1(cpt.init, DATA, zeta.group, w = 1/3, ...)
}
\arguments{
\item{cpt.init}{A \code{integer} vector of initial changepoints estimation (sorted in strictly increasing order).}

\item{DATA}{A \code{numeric} matrix of observations with with horizontal axis being time, and vertical axis being dimensions.}

\item{zeta.group}{A \code{numeric} scalar of lasso penalty.}

\item{w}{A \code{numeric} scalar of weight for interpolation.}

\item{...}{Additional arguments.}
}
\value{
An \code{integer} vector of locally refined change points estimation.
}
\description{
Perform local refinement for VAR1 change points detection.
}
\examples{
p = 10
sigma = 1
n = 5
v1 = 2*(seq(1,p,1)\%\%2) - 1
v2 = -v1
AA = matrix(0, nrow = p, ncol = p-2)
A1=cbind(v1,v2,AA)
A2=cbind(v2,v1,AA)
A3=A1
data = simu.VAR1(sigma, p, 2*n+1, A1)
data = cbind(data, simu.VAR1(sigma, p, 2*n, A2, vzero=c(data[,ncol(data)])))
data = cbind(data, simu.VAR1(sigma, p, 2*n, A3, vzero=c(data[,ncol(data)])))
N = ncol(data)
X_curr = data[,1:(N-1)]
X_futu = data[,2:N]
parti = DP.VAR1(X_futu, X_curr, gamma = 1, lambda = 1, delta = 5)$partition
cpt.init = part2local(parti)
local.refine.VAR1(cpt.init, data, 1, 1/3)
}
\author{

}
