% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_nonparametric.R
\name{MNWBS}
\alias{MNWBS}
\title{Wild binary segmentation for multivariate nonparametric change points detection}
\usage{
MNWBS(Y, s, e, Alpha, Beta, h, level = 0, ...)
}
\arguments{
\item{Y}{A \code{numeric} matrix of observations with horizontal axis being time, and vertical axis being multiple observations on each time point.}

\item{s}{A \code{integer} scalar of starting index.}

\item{e}{A \code{integer} scalar of ending index.}

\item{Alpha}{A \code{integer} vector of starting indices of random intervals.}

\item{Beta}{A \code{integer} vector of ending indices of random intervals.}

\item{h}{A \code{numeric} vector bandwidth parameter}

\item{level}{Should be fixed as 0.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item S           A vector of estimated changepoints (sorted in strictly increasing order).
 \item Dval        A vector of values of CUSUM statistic based on KS distance.
 \item Level       A vector representing the levels at which each change point is detected.
 \item Parent      A matrix with the starting indices on the first row and the ending indices on the second row.
 \item ...         Additional parameters.
}
}
\description{
TO DO
}
\examples{
new_MWBS

}
\author{
Haotian Xu
}
