% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_nonparametric.R
\name{NBS}
\alias{NBS}
\title{Standard binary segmentation for univariate nonparametric change points detection}
\usage{
NBS(Y, delta, s, e, N, ...)
}
\arguments{
\item{Y}{A \code{numeric} matrix of observations with with horizontal axis being time, and vertical axis being multiple observations on each time point.}

\item{delta}{A strictly \code{integer} scalar of minimum spacing.}

\item{s}{A \code{integer} scalar of starting index.}

\item{e}{A \code{integer} scalar of ending index.}

\item{N}{A \code{integer} vector representing number of multiple observations on each time point.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item S           A vector of estimated changepoints (sorted in strictly increasing order).
 \item Dval        A vector of values of CUSUM statistic based on KS distance.
 \item ...         Additional parameters.
}
}
\description{
TO DO
}
\examples{
plot.ts(t(Y))
points(x = tail(temp$S[order(temp$Dval)],20), y =rep(0, 20), col = "red")
}
\author{
Haotian Xu
}
