% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{online.network}
\alias{online.network}
\title{Online changepoint detection for network data by controlling the false alarm rate at level alpha.}
\usage{
online.network(
  data_mat1,
  data_mat2,
  b_vec,
  C = 32.1 * (2^0.25) * exp(2),
  rho,
  alpha,
  ...
)
}
\arguments{
\item{data_mat1}{A \code{numeric} matrix of observations with with horizontal axis being time, and with each column be the vectorized adjacency matrix.}

\item{data_mat2}{A \code{numeric} matrix of observations with with horizontal axis being time, and with each column be the vectorized adjacency matrix (data_mat1 and data_mat2 are independent and have the same dimensions ).}

\item{b_vec}{A \code{numeric} vector of thresholds b_t with t >= 2.}

\item{C}{A \code{numeric} scalar of tuning constant corresponding to the threshold for the frobenius norm of the USVT estimator.}

\item{rho}{A \code{numeric} scalar of sparsity parameter of the network data.}

\item{alpha}{A \code{numeric} scalar in (0,1) representing the level.}

\item{...}{Additional arguments.}
}
\value{
A \code{list} with the structure:
\itemize{
 \item t:           Estimated changepoint.
 \item score:       A \code{numeric} vector of computed cumsum statistics.
 \item b_vec        A \code{numeric} vector of thresholds b_t with t >= 2.
}
}
\description{
Perform online changepoint detection for network data by controlling the false alarm rate at level alpha.
}
