% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_partitioning.R
\name{optimal_partitioning}
\alias{optimal_partitioning}
\title{Optimal Partitioning Method}
\usage{
optimal_partitioning(x, beta)
}
\arguments{
\item{x}{A numeric vector representing the data to segment.}

\item{beta}{A double value representing the penalty term for adding a new segment.}
}
\value{
A list with (1) the changepoint elements (each last index of each segment in \code{changepoints}), (2) a vector `\code{nb} saving the number of non-pruned elements at each iteration, (3) a vector \code{lastIndexSet} containing the non-pruned indices at the end of the algo and (4) a vector \code{costQ} saving the optimal cost at each time step.
}
\description{
This function computes the optimal partitioning of a given vector x with a given penalty term beta. It finds the optimal changepoints that minimize the cost function using dynamic programming.
}
\details{
Optimal Partitioning Method using R
}
