% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_partitioning.R
\name{optimal_partitioning}
\alias{optimal_partitioning}
\title{Optimal Partitioning Method}
\usage{
optimal_partitioning(x, beta)
}
\arguments{
\item{x}{A numeric vector representing the data to segment.}

\item{beta}{A double value representing the penalty term for adding a new segment.}
}
\value{
A vector of indices representing the optimal breakpoints.
}
\description{
This function computes the optimal partitioning of a given vector x with a given penalty term beta. It finds the optimal changepoints that minimize the cost function using dynamic programming.
}
\details{
Optimal Partitioning Method using R
}
