% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{part2local}
\alias{part2local}
\title{Partition to localization.}
\usage{
part2local(parti_vec)
}
\arguments{
\item{parti_vec}{A \code{numeric} vector of observations.}

\item{...}{Additional arguments.}
}
\value{
A vector of change point locations.
}
\description{
Find change point locations from the best partition produced by dynamic programming.
}
\examples{
y = c(rep(0, 100), rep(1, 100))
part2local(DP.univar(y, 1, 4)$partition)
}
\author{
Haotian Xu
}
