% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{part2local}
\alias{part2local}
\title{Partition to localization}
\usage{
part2local(x)
}
\arguments{
\item{x}{A \code{numeric} vector of observations.}

\item{...}{Additional arguments.}
}
\value{
TO DO.
}
\description{
TO DO
}
\examples{
TO DO
}
\author{
Haotian Xu
}
