% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R
\name{pelt}
\alias{pelt}
\title{Optimal Partitioning using PELT}
\usage{
pelt(data, penalty)
}
\arguments{
\item{data}{A numeric vector representing the data to segment.}

\item{penalty}{A double value representing the penalty term for adding a new segment.}
}
\value{
A list with (1) the changepoint elements (each last index of each segment in \code{changepoints}), (2) a vector `\code{nb} saving the number of non-pruned elements at each iteration, (3) a vector \code{lastIndexSet} containing the non-pruned indices at the end of the algo and (4) a vector \code{costQ} saving the optimal cost at each time step.
}
\description{
This function implements the OP algorithm using PELT of a given vector `data` with a given penalty term.
It finds the optimal changepoints that minimize the cost function using dynamic programming.
}
\details{
Optimal Partitioning algorithm using PELT
}
\examples{
n <- 1000
data <- rep(c(0, 5, 2.5, 7), each = n) + rnorm(4 * n)
penalty <- 2 * log(length(penalty))
PELT_R <- pelt(data, penalty)
plot_segmentation(data, PELT_R$changepoints, title = "PELT Segmentation")

}
