% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R
\name{pelt}
\alias{pelt}
\title{PELT Method}
\usage{
pelt(x, penalty, minseglen = 1)
}
\arguments{
\item{x}{A numeric vector representing the data to segment.}

\item{penalty}{A double value representing the penalty term for adding a new segment.}

\item{minseglen}{The minimum length of the segment}
}
\value{
An integer vector of indices representing the optimal breakpoints.
}
\description{
This function implements the PELT (Pruned Exact Linear Time) method in pure R.
It computes the optimal partitioning of a given numeric vector `x` using a penalized cost function.
}
\details{
PELT Method using R
}
\examples{
x <- c(1, 2, 3, 10, 10, 10, 20, 20, 20, 5, 5, 5)
penalty <- 5
pelt(x, penalty)

}
