% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_pelt}
\alias{rcpp_pelt}
\title{PELT Method}
\usage{
rcpp_pelt(x, penalty, minseglen = 1L)
}
\arguments{
\item{x}{A numeric vector representing the data to segment.}

\item{penalty}{A double value representing the penalty term for adding a new segment.}

\item{minseglen}{The minimum length of the segment.}
}
\value{
An integer vector of indices representing the optimal breakpoints.
}
\description{
This function implements the PELT (Pruned Exact Linear Time) method in Rcpp.
}
\details{
PELT Method using C++
}
