% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{simu.SBM}
\alias{simu.SBM}
\title{Simulate a Stochastic Block Model (without change point).}
\usage{
simu.SBM(connec_mat, can_vec, n, ...)
}
\arguments{
\item{connec_mat}{A \code{integer} matrix representing the connectivity matrix (symmetric matrix with dimension being the number of blocks, and entries being 0 or 1).}

\item{can_vec}{A \code{integer} p-dim vector representing the candidate vector.}

\item{n}{A \code{integer} scalar representing the number of observations.}

\item{...}{Additional arguments.}
}
\value{
A p^2-by-n matrix.
}
\description{
Simulate a Stochastic Block Model (without change point).
}
\examples{
TO DO
}
\author{

}
