% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEPP.R
\name{simu.SEPP}
\alias{simu.SEPP}
\title{Simulate a (stable) SEPP model (without change point).}
\usage{
simu.SEPP(intercept, M, A, threshold, n, vzero = NULL, ...)
}
\arguments{
\item{intercept}{A \code{numeric} scalar representing the intercept of the model.}

\item{M}{An \code{integer} scalar representing dimensionality.}

\item{A}{A \code{numeric} M-by_M matrix representing the coefficient matrix.}

\item{threshold}{A \code{numeric} scalar representing the upper bound for each coordinate of X_t (for stability).}

\item{n}{An \code{integer} scalar representing sample size.}

\item{vzero}{A \code{numeric} vector representing the observation at time 0.}

\item{...}{Additional arguments.}
}
\value{
A p-by-n matrix.
}
\description{
Simulate a (stable) SEPP model (without change point).
}
\examples{
M = 30
n = 300
threshold = 6 #thresholding makes the process stable
intercept = 1/2 #intercept of the model. Assume to be known as in the existing literature
v1 = (2*(seq(1,30,1)\%\%2)-1)*(0.05+0.05*6)
v2 = -v1
AA=matrix(0,nrow = 30,ncol=28)
A=cbind(v1,v2,AA)
A2=cbind(v2,v1,AA)
DATA1 = simu.SEPP(intercept, M, A, threshold, n, vzero = NULL)
DATA2 = simu.SEPP(intercept, M, A2, threshold, n, vzero = DATA1[,n])
DATA = cbind(DATA1, DATA2)
}
\references{
Wang, D., Yu, Y., & Willett, R. (2020). Detecting Abrupt Changes in High-Dimensional Self-Exciting Poisson Processes. arXiv preprint arXiv:2006.03572.
}
\author{
Daren Wang & Haotian Xu
}
