% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVP_ALGO_VALIDITYtests.R
\name{valid_focus}
\alias{valid_focus}
\title{Validity Test Based on FOCuS}
\usage{
valid_focus(y, gamma)
}
\arguments{
\item{y}{A numeric vector representing a segment of the signal.}

\item{gamma}{A numeric threshold passed to the FOCuS method.}
}
\value{
A logical value indicating whether the segment is considered valid (TRUE if no changepoint detected).
}
\description{
Checks whether a given segment is valid using the FOCuS method (i.e., has no changepoint).
}
\details{
Validity Test Based on FOCuS
}
