% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVP_ALGO_VALIDITYtests.R
\name{valid_Range_slack}
\alias{valid_Range_slack}
\title{Validity Test Based on Trimmed Range}
\usage{
valid_Range_slack(y, gamma)
}
\arguments{
\item{y}{A numeric vector representing a segment of the signal.}

\item{gamma}{A numeric threshold for the trimmed range.}
}
\value{
TRUE if the trimmed range is less than or equal to gamma, or if the segment is too short.
}
\description{
Applies a slack range test using trimmed minimum and maximum (ignores smallest and largest 3 values).
}
\details{
Validity Test Based on Trimmed Range
}
